ai-services/requirements.txt
# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0

# Machine Learning and AI
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
transformers==4.36.2
torch==2.1.2
textblob==0.17.1

# OpenAI Integration
openai==1.6.1

# Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0

# Database connectivity
pymongo==4.6.1
redis==5.0.1

# API and serialization
marshmallow==3.20.2
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0

# Development tools
pytest==7.4.3
black==23.12.1
flake8==6.1.0
